test(ui): use real VNC for Desktop proof
